Context

Class representing a PetroVisor context object.

Public fields

name

The name of the context.

entity_set

The context's entity set (object of class EntitySet).

scope

The context's scope (object of class Scope).

hierarchy

(Optional) The hierarchy used for automatic aggregation (Object of type Hierarchy).

loading_scenario_name

(Optional) The name of the loading scenario.

saving_scenario_name

(Optional) The name of the saving scenario.

scenario_data_only

(Optional) Whether to load data from the specified loading scenario only. If FALSE, data will be merged with workspace data.

formula

The context's definition as string (P# syntax).

description

The description of the item.

labels

A list of strings holding the labels of the context.

Methods

Public methods

Context$new()

Context$toList()

Context$clone()

Method new()

Create a new Context instance.

Usage

Context$new(
  name = NULL,
  entity_set = NULL,
  scope = NULL,
  hierarchy = NULL,
  loading_scenario_name = NULL,
  saving_scenario_name = NULL,
  scenario_data_only = FALSE,
  formula = NULL,
  description = NULL,
  labels = list()
)

Arguments

name

The name of the context.

entity_set

The context's entity set (object of class EntitySet).

scope

The context's scope (object of class Scope).

hierarchy

(Optional) The hierarchy used for automatic aggregation (Object of type Hierarchy).

loading_scenario_name

(Optional) The name of the loading scenario.

saving_scenario_name

(Optional) The name of the saving scenario.

scenario_data_only

(Optional) Whether to load data from the specified loading scenario only. If FALSE, data will be merged with workspace data.

formula

The context's definition as string (P# syntax).

description

The description of the item.

labels

A list of strings holding the labels of the context.

Method toList()

Usage

Context$toList()

Details

Convert the object to a list. This function is mainly used by the RepositoryService to convert the objects to lists and then call the web API.

Method clone()

The objects of this class are cloneable with this method.

Usage

Context$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples